# server/src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is API documentation for Redland Museum's DIGS administration portal",
    "version": "1.0.0",
    "title": "Swagger Redland Museum",
    "contact": {
      "email": "tyronewessnolasco@gmail.com"
    }
  },
  "host": "rm.kongroo.xyz",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "paths": {
...
    "/artefacts": {
      "get": {
        "tags": [
          "artefacts"
        ],
        "summary": "List all artefacts",
        "description": "",
        "operationId": "listArtefacts",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "404": {
            "description": "Artefacts not found"
          }
        }
      },
      "post": {
        "tags": [
          "artefacts"
        ],
        "summary": "Add a new artefact to the list",
        "description": "",
        "operationId": "addArtefacts",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Artefact that needs to be added to the list",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Artefact"
            }
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/artefacts/{id}": {
      "get": {
        "tags": [
          "artefacts"
        ],
        "summary": "Find artefact by ID",
        "description": "Returns a single artefact",
        "operationId": "getArtefactById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of artefact to return",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Artefact"
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Artefact not found"
          }
        }
      },
      "put": {
        "tags": [
          "artefacts"
        ],
        "summary": "Updates a artefact with form data",
        "description": "Updates artefact",
        "operationId": "updateArtefactWithForm",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "formData",
            "description": "ID of the to be updated artefact",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "name",
            "in": "formData",
            "description": "Updated name of the artefact",
            "required": true,
            "type": "string"
          },
          {
            "name": "description",
            "in": "formData",
            "description": "Updated description of the artefact",
            "required": false,
            "type": "string"
          },
          {
            "name": "acquisitionDate",
            "in": "formData",
            "description": "Updated acquisition date of the artefact",
            "required": false,
            "type": "string",
            "format": "date-time"
          },
          {
            "name": "createdAt",
            "in": "formData",
            "description": "Updated create date of the artefact entry",
            "required": false,
            "type": "string",
            "format": "date-time"
          },
          {
            "name": "updatedAt",
            "in": "formData",
            "description": "Updated update date of artefact entry",
            "required": false,
            "type": "string",
            "format": "date-time"
          },
          {
            "name": "thumbnail",
            "in": "formData",
            "description": "Updated image of the artefact in base64",
            "required": false,
            "type": "string",
            "format": "base64"
          },
          {
            "name": "coordX",
            "in": "formData",
            "description": "Updated x coordinate of the artefact in base64",
            "required": false,
            "type": "integer",
            "format": "base64"
          },
          {
            "name": "coordY",
            "in": "formData",
            "description": "Updated y coordinate of the artefact in base64",
            "required": false,
            "type": "integer",
            "format": "base64"
          },
          {
            "name": "zone",
            "in": "formData",
            "description": "Updated zone ID of artefact",
            "required": false,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "artefacts"
        ],
        "summary": "Deletes a artefact",
        "description": "",
        "operationId": "deleteArtefact",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Artefact id to delete",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Artefact not found"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
    "/exhibitions": {
      "get": {
        "tags": [
          "exhibitions"
        ],
        "summary": "List all exhibitions",
        "description": "",
        "operationId": "listExhibitions",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "404": {
            "description": "Exhibitions not found"
          }
        }
      },
      "post": {
        "tags": [
          "exhibitions"
        ],
        "summary": "Add a new exhibition to the list",
        "description": "",
        "operationId": "addExhibitions",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Exhibition that needs to be added to the list",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Exhibition"
            }
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/exhibitions/{id}": {
      "get": {
        "tags": [
          "exhibitions"
        ],
        "summary": "Find exhibition by ID",
        "description": "Returns a single exhibition",
        "operationId": "getExhibitionById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of exhibition to return",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Exhibition"
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Exhibition not found"
          }
        }
      },
      "put": {
        "tags": [
          "exhibitions"
        ],
        "summary": "Updates a exhibition with form data",
        "description": "",
        "operationId": "updateExhibitionsWithForm",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "formData",
            "description": "ID of the to be updated exhibition",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "name",
            "in": "formData",
            "description": "Updated name of the exhibition",
            "required": true,
            "type": "string"
          },
          {
            "name": "description",
            "in": "formData",
            "description": "Updated description of the exhibition",
            "required": false,
            "type": "string"
          },
          {
            "name": "organiser",
            "in": "formData",
            "description": "Updated organiser of the exhibition",
            "required": false,
            "type": "string",
            "format": "date-time"
          },
          {
            "name": "priceAdult",
            "in": "formData",
            "description": "Updated adult price of the exhibition",
            "required": false,
            "type": "integer",
            "format": "double"
          },
          {
            "name": "priceConcession",
            "in": "formData",
            "description": "Updated concession price of the exhibition",
            "required": false,
            "type": "integer",
            "format": "double"
          },
          {
            "name": "priceChild",
            "in": "formData",
            "description": "Updated child price of the exhibition",
            "required": false,
            "type": "integer",
            "format": "double"
          },
          {
            "name": "startDate",
            "in": "formData",
            "description": "Updated start date of the exhibition",
            "required": false,
            "type": "string",
            "format": "date-time"
          },
          {
            "name": "finishDate",
            "in": "formData",
            "description": "Updated finish date of exhibition",
            "required": false,
            "type": "string",
            "format": "date-time"
          },
          {
            "name": "createdAt",
            "in": "formData",
            "description": "Updated create date of the exhibition entry",
            "required": false,
            "type": "string",
            "format": "date-time"
          },
          {
            "name": "updatedAt",
            "in": "formData",
            "description": "New update date of exhibition entry",
            "required": false,
            "type": "string",
            "format": "date-time"
          },
          {
            "name": "thumbnail",
            "in": "formData",
            "description": "Updated thumbnail of the exhibition in base64",
            "required": false,
            "type": "string",
            "format": "base64"
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "exhibitions"
        ],
        "summary": "Deletes a exhibition",
        "description": "",
        "operationId": "deleteExhibition",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Exhibition id to delete",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Exhibition not found"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "definitions": {
    "Artefact": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "acquisitionDate": {
          "type": "string",
          "format": "date-time"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        },
        "coordX": {
          "type": "number"
        },
        "coordY": {
          "type": "number"
        },
        "zoneId": {
          "type": "number"
        },
        "thumbnail": {
          "type": "string",
          "format": "base64"
        }
      },
      "xml": {
        "name": "Artefact"
      }
    },
...
    "Exhibition": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "organiser": {
          "type": "string"
        },
        "priceAdult": {
          "type": "integer",
          "format": "double"
        },
        "priceConcession": {
          "type": "integer",
          "format": "double"
        },
        "priceChild": {
          "type": "integer",
          "format": "double"
        },
        "startDate": {
          "type": "string",
          "format": "date-time"
        },
        "finishDate": {
          "type": "string",
          "format": "date-time"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        },
        "thumnail": {
          "type": "string",
          "format": "base64"
        }
      },
      "xml": {
        "name": "Exhibition"
      }
    },
...
  }
}